requirements.txt
-r requirements-zero.txt
-r requirements-account.txt
-r requirements-dev.txt

dj-database-url==0.4.2
psycopg2==2.7.3.2
gunicorn==19.7.1
whitenoise==3.3.1
raven==6.4.0
pyuca==1.2
flake8==3.5.0
django-cors-headers==2.1.0
django-querycount==0.7.0
mozilla_django_oidc==0.5.0
